Move serializer under utils and add  the deserialization before settings product data to index model
