+++ Model/Service/Index.php
@@ -60,7 +60,7 @@ use Nosto\Tagging\Model\ResourceModel\Magento\Product\Collection as ProductColle
 use Nosto\Tagging\Model\ResourceModel\Magento\Product\CollectionFactory as ProductCollectionFactory;
 use Nosto\Tagging\Model\ResourceModel\Product\Index\Collection as NostoIndexCollection;
 use Nosto\Tagging\Model\ResourceModel\Product\Index\CollectionFactory as NostoIndexCollectionFactory;
-use Nosto\Tagging\Model\Service\Serializer\SerializedProductBuilder;
+use Nosto\Tagging\Util\Serializer\ProductSerializer;
 use Nosto\Tagging\Util\Iterator;
 use Nosto\Tagging\Util\Product as ProductUtil;
 use Nosto\Types\Product\ProductInterface as NostoProductInterface;
@@ -110,8 +110,8 @@ class Index extends AbstractService
     /** @var array */
     private $invalidatedProducts = [];
 
-    /** @var SerializedProductBuilder */
-    private $serializedProductBuilder;
+    /** @var ProductSerializer */
+    private $productSerializer;
 
     /**
      * Index constructor.
@@ -128,7 +128,7 @@ class Index extends AbstractService
      * @param TimezoneInterface $magentoTimeZone
      * @param NostoDataHelper $nostoDataHelper
      * @param Sync $nostoSyncService
-     * @param SerializedProductBuilder $serializedProductBuilder
+     * @param ProductSerializer $productSerializer
      */
     public function __construct(
         IndexRepository $indexRepository,
@@ -144,7 +144,7 @@ class Index extends AbstractService
         TimezoneInterface $magentoTimeZone,
         NostoDataHelper $nostoDataHelper,
         Sync $nostoSyncService,
-        SerializedProductBuilder $serializedProductBuilder
+        ProductSerializer $productSerializer
     ) {
         parent::__construct($nostoDataHelper, $logger);
         $this->indexRepository = $indexRepository;
@@ -158,7 +158,7 @@ class Index extends AbstractService
         $this->productCollectionFactory = $productCollectionFactory;
         $this->magentoTimeZone = $magentoTimeZone;
         $this->nostoSyncService = $nostoSyncService;
-        $this->serializedProductBuilder = $serializedProductBuilder;
+        $this->productSerializer = $productSerializer;
     }
 
     /**
@@ -318,7 +318,7 @@ class Index extends AbstractService
             );
             $store = $this->nostoHelperScope->getStore($productIndex->getStoreId());
             $nostoProduct = $this->nostoProductBuilder->build($magentoProduct, $store);
-            $nostoIndexedProduct = $this->serializedProductBuilder->fromString(
+            $nostoIndexedProduct = $this->productSerializer->fromString(
                 $productIndex->getProductData()
             );
             if ($nostoIndexedProduct instanceof NostoProductInterface === false ||
@@ -327,7 +327,11 @@ class Index extends AbstractService
                     && !ProductUtil::isEqual($nostoProduct, $nostoIndexedProduct)
                 )
             ) {
-                $productIndex->setProductData($nostoProduct);
+                $productIndex->setProductData(
+                    $this->productSerializer->toString(
+                        $nostoProduct
+                    )
+                );
                 $productIndex->setInSync(false);
             }
             $productIndex->setIsDirty(false);
